refactor: account_spot to spot v1
