+++ integrationtest/testdata/spot-v1/spot-v1-delete-order-fast.json
