+++ integrationtest/testdata/spot-v1/spot-v1-order-batch-cancel.json
